+++ src/str_term.rs
@@ -1,46 +1,46 @@
 pub(crate) mod str_types {
-    pub(crate) const STR_FUNC_ESCAPE: usize = 0x01;
-    pub(crate) const STR_FUNC_EXPAND: usize = 0x02;
-    pub(crate) const STR_FUNC_REGEXP: usize = 0x04;
-    pub(crate) const STR_FUNC_QWORDS: usize = 0x08;
-    pub(crate) const STR_FUNC_SYMBOL: usize = 0x10;
-    pub(crate) const STR_FUNC_INDENT: usize = 0x20;
-    pub(crate) const STR_FUNC_LABEL: usize = 0x40;
-    pub(crate) const STR_FUNC_LIST: usize = 0x4000;
-    pub(crate) const STR_FUNC_TERM: usize = 0x8000;
+    pub(crate) const STR_FUNC_ESCAPE: u32 = 0x01;
+    pub(crate) const STR_FUNC_EXPAND: u32 = 0x02;
+    pub(crate) const STR_FUNC_REGEXP: u32 = 0x04;
+    pub(crate) const STR_FUNC_QWORDS: u32 = 0x08;
+    pub(crate) const STR_FUNC_SYMBOL: u32 = 0x10;
+    pub(crate) const STR_FUNC_INDENT: u32 = 0x20;
+    pub(crate) const STR_FUNC_LABEL: u32 = 0x40;
+    pub(crate) const STR_FUNC_LIST: u32 = 0x4000;
+    pub(crate) const STR_FUNC_TERM: u32 = 0x8000;
 
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_label: usize = STR_FUNC_LABEL;
+    pub(crate) const str_label: u32 = STR_FUNC_LABEL;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_squote: usize = 0;
+    pub(crate) const str_squote: u32 = 0;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_dquote: usize = STR_FUNC_EXPAND;
+    pub(crate) const str_dquote: u32 = STR_FUNC_EXPAND;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_xquote: usize = STR_FUNC_EXPAND;
+    pub(crate) const str_xquote: u32 = STR_FUNC_EXPAND;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_regexp: usize = STR_FUNC_REGEXP | STR_FUNC_ESCAPE | STR_FUNC_EXPAND;
+    pub(crate) const str_regexp: u32 = STR_FUNC_REGEXP | STR_FUNC_ESCAPE | STR_FUNC_EXPAND;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_sword: usize = STR_FUNC_QWORDS | STR_FUNC_LIST;
+    pub(crate) const str_sword: u32 = STR_FUNC_QWORDS | STR_FUNC_LIST;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_dword: usize = STR_FUNC_QWORDS | STR_FUNC_EXPAND | STR_FUNC_LIST;
+    pub(crate) const str_dword: u32 = STR_FUNC_QWORDS | STR_FUNC_EXPAND | STR_FUNC_LIST;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_ssym: usize = STR_FUNC_SYMBOL;
+    pub(crate) const str_ssym: u32 = STR_FUNC_SYMBOL;
     #[allow(non_upper_case_globals)]
-    pub(crate) const str_dsym: usize = STR_FUNC_SYMBOL | STR_FUNC_EXPAND;
+    pub(crate) const str_dsym: u32 = STR_FUNC_SYMBOL | STR_FUNC_EXPAND;
 }
 
 #[derive(Debug, Clone, Default)]
 pub(crate) struct HeredocEnd<'b> {
-    pub(crate) start: usize,
-    pub(crate) end: usize,
+    pub(crate) start: u32,
+    pub(crate) end: u32,
     pub(crate) value: &'b [u8],
 }
 
 #[derive(Debug, Clone, Default)]
 pub(crate) struct StringLiteral<'b> {
     // struct rb_strterm_literal_struct
-    pub(crate) nest: usize,
-    pub(crate) func: usize,
+    pub(crate) nest: u32,
+    pub(crate) func: u32,
     pub(crate) paren: Option<u8>,
     pub(crate) term: u8,
     pub(crate) heredoc_end: Option<HeredocEnd<'b>>,
@@ -48,8 +48,8 @@ pub(crate) struct StringLiteral<'b> {
 
 impl<'b> StringLiteral<'b> {
     pub(crate) fn new(
-        nest: usize,
-        func: usize,
+        nest: u32,
+        func: u32,
         paren: Option<u8>,
         term: u8,
         heredoc_end: Option<HeredocEnd<'b>>,
@@ -66,23 +66,23 @@ impl<'b> StringLiteral<'b> {
 
 #[derive(Debug, Clone, Default)]
 pub(crate) struct HeredocLiteral {
-    pub(crate) lastline: usize, /* the string of line that contains `<<"END"` */
-    pub(crate) offset: usize,   /* the column of END in `<<"END"` */
-    pub(crate) sourceline: usize, /* lineno of the line that contains `<<"END"` */
-    pub(crate) length: usize,   /* the length of END in `<<"END"` */
+    pub(crate) lastline: u32,   /* the string of line that contains `<<"END"` */
+    pub(crate) offset: u32,     /* the column of END in `<<"END"` */
+    pub(crate) sourceline: u32, /* lineno of the line that contains `<<"END"` */
+    pub(crate) length: u32,     /* the length of END in `<<"END"` */
 
-    pub(crate) quote: usize,
-    pub(crate) func: usize,
+    pub(crate) quote: u32,
+    pub(crate) func: u32,
 }
 
 impl HeredocLiteral {
     pub(crate) fn new(
-        lastline: usize,
-        offset: usize,
-        sourceline: usize,
-        length: usize,
-        quote: usize,
-        func: usize,
+        lastline: u32,
+        offset: u32,
+        sourceline: u32,
+        length: u32,
+        quote: u32,
+        func: u32,
     ) -> Self {
         Self {
             lastline,
